Core module separated. GitHub Extension module introduced (to be implemented).
The App module contains the main app and depends on Core.

The Core library module contains core functionality as is used in other modules.

The extension modules contain extension apps which depend on Core,
have no standalone UI and are used by the main App (to be inplemented).

The GitHub Extension module is experimental.
It is supposed to introduce a GitHub Type in addition to included Blog Type,
but as a separately installed app.

Third party modules would be parts of third party projects, depending on included Core module.